3. auth usecase 를 user usecase로 통합
