+++ src/main/java/com/nexters/phochak/auth/AuthAspect.java
@@ -3,8 +3,8 @@ package com.nexters.phochak.auth;
 import com.nexters.phochak.common.exception.PhochakException;
 import com.nexters.phochak.common.exception.ResCode;
 import com.nexters.phochak.user.application.JwtTokenService;
-import com.nexters.phochak.user.application.UserService;
 import com.nexters.phochak.user.application.port.in.JwtTokenUseCase;
+import com.nexters.phochak.user.application.port.in.UserUseCase;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -25,7 +25,7 @@ public class AuthAspect {
     public static final String WHITE_SPACE = " ";
     private final HttpServletRequest httpServletRequest;
     private final JwtTokenUseCase jwtTokenUseCase;
-    private final UserService userService;
+    private final UserUseCase userService;
 
     @Around("@annotation(com.nexters.phochak.auth.Auth)")
     public Object validateAccessToken(final ProceedingJoinPoint joinPoint) throws Throwable {
