--- src/main/java/com/nexters/phochak/user/application/port/in/AuthUseCase.java
@@ -1,11 +1,0 @@
-package com.nexters.phochak.user.application.port.in;
-
-public interface AuthUseCase {
-    /**
-     * OAuth 로그인을 진행한다.
-     *
-     * @param provider
-     * @param requestDto
-     */
-    Long login(String provider, LoginRequestDto requestDto);
-}
